--- pytorch_forecasting/layers/_m_lstm/__init__.py
@@ -1,7 +1,0 @@
-"""mLSTM layer"""
-
-from pytorch_forecasting.layers._m_lstm.cell import mLSTMCell
-from pytorch_forecasting.layers._m_lstm.layer import mLSTMLayer
-from pytorch_forecasting.layers._m_lstm.network import mLSTMNetwork
-
-__all__ = ["mLSTMCell", "mLSTMLayer", "mLSTMNetwork"]
