+++ pytorch_forecasting/layers/_slstm/__init__.py
@@ -1,0 +1,7 @@
+"""sLSTM layer"""
+
+from pytorch_forecasting.layers._slstm.cell import sLSTMCell
+from pytorch_forecasting.layers._slstm.layer import sLSTMLayer
+from pytorch_forecasting.layers._slstm.network import sLSTMNetwork
+
+__all__ = ["sLSTMCell", "sLSTMLayer", "sLSTMNetwork"]
