+++ pytorch_forecasting/models/__init__.py
@@ -20,7 +20,7 @@ from pytorch_forecasting.models.temporal_fusion_transformer import (
 )
 from pytorch_forecasting.models.tide import TiDEModel
 from pytorch_forecasting.models.timexer import TimeXer
-from pytorch_forecasting.models.x_lstm_time import xLSTMTime
+from pytorch_forecasting.models.xlstm import xLSTMTime
 
 __all__ = [
     "NBeats",
