--- pytorch_forecasting/models/x_lstm_time/__init__.py
@@ -1,6 +1,0 @@
-"""xLSTMTime implementation for forecasting."""
-
-from pytorch_forecasting.models.x_lstm_time._xlstm_pkg import xLSTMTime_pkg
-from pytorch_forecasting.models.x_lstm_time.x_lstm import xLSTMTime
-
-__all__ = ["xLSTMTime", "xLSTMTime_pkg"]
