+++ pytorch_forecasting/models/xlstm/__init__.py
@@ -1,0 +1,6 @@
+"""xLSTMTime implementation for forecasting."""
+
+from pytorch_forecasting.models.xlstm._xlstm import xLSTMTime
+from pytorch_forecasting.models.xlstm._xlstm_pkg import xLSTMTime_pkg
+
+__all__ = ["xLSTMTime", "xLSTMTime_pkg"]
